requirements.txt
# Google API Libraries
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
python-dateutil==2.9.0
validators==0.28.1

# Retry and backoff
backoff==2.2.1
tenacity==8.2.3

# CLI
typer==0.12.3
rich==13.7.1

# Data handling
pandas==3.0.0rc0

# Testing
pytest==8.4.2
pytest-mock==3.15.1
protobuf~=6.33.2

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1